requirements.txt
# requirements.txt

composio-openai>=0.1.0
openai>=0.27.0

# Websocket server support
websockets>=12.0

# Env file loading
python-dotenv>=0.21.0

# For Twitter OAuth in skill_x_api.py
requests>=2.28.0
requests_oauthlib>=1.3.1

# Testing
pytest
pytest-asyncio
requests
aiohttp
asyncio

# Market Research
linkup-sdk>=0.1.0

# Rich Console Output
rich>=13.7.0

# Version pinning to resolve conflicts
protobuf>=4.21.8,<5.0.0

# Remove conflicting dependencies and packages
# Removed: replit, opentelemetry-proto
# These packages have conflicting protobuf version requirements